third_party/WebKit/LayoutTests/inspector/elements/styles/internal-properties-text-expected.txt

Tests that querying -internal-* presentation attribute from the inspector works as expected.

Inspect Me
=== Rules for <marquee> ===
[expanded] 
element.style  { ()

[expanded] 
marquee[Attributes Style]  { ()
